requirements.txt
django
django-environ
google-genai